Moved debug code to specific package, no longer runs in built versions either
