+++ gradle.properties
@@ -10,6 +10,6 @@ loader_version=0.14.19-babric.1-bta
 halplibe_version=2.7.0
 
 # Mod
-mod_version=0.1.1
+mod_version=0.2.0
 mod_group=useless
 mod_name=dragonfly
