+++ src/main/java/useless/dragonfly/debug/testentity/RenderZombieTest.java
@@ -1,4 +1,4 @@
-package useless.dragonfly.testentity;
+package useless.dragonfly.debug.testentity;
 
 import net.minecraft.client.render.entity.LivingRenderer;
 import net.minecraft.client.render.model.ModelBase;
