+++ src/main/java/useless/dragonfly/debug/testentity/ZombieModelTest.java
@@ -1,4 +1,4 @@
-package useless.dragonfly.testentity;
+package useless.dragonfly.debug.testentity;
 
 import net.minecraft.core.util.helper.MathHelper;
 import useless.dragonfly.model.entity.BenchEntityModel;
