+++ src/main/java/useless/dragonfly/mixins/mixin/MinecraftMixin.java
@@ -1,4 +1,4 @@
-package useless.dragonfly.mixin;
+package useless.dragonfly.mixins.mixin;
 
 import net.minecraft.client.Minecraft;
 import org.spongepowered.asm.mixin.Mixin;
