+++ src/main/java/useless/dragonfly/model/block/data/CubeData.java
@@ -2,6 +2,7 @@ package useless.dragonfly.model.block.data;
 
 import com.google.gson.annotations.SerializedName;
 import useless.dragonfly.DragonFly;
+import useless.dragonfly.utilities.Utilities;
 
 import java.util.Arrays;
 import java.util.HashMap;
@@ -24,13 +25,13 @@ public class CubeData {
 		builder.append("from: ").append(Arrays.toString(from)).append("\n");
 		builder.append("to: ").append(Arrays.toString(to)).append("\n");
 		builder.append("rotation: ").append("\n");
-		builder.append(DragonFly.tabBlock(rotation.toString(), 1));
+		builder.append(Utilities.tabBlock(rotation.toString(), 1));
 		builder.append("shade: ").append(shade).append("\n");
 		builder.append("color: ").append(color).append("\n");
 		builder.append("faces: ").append("\n");
 		for (String key: faces.keySet()) {
 			builder.append("\t").append(key).append(": \n");
-			builder.append(DragonFly.tabBlock(faces.get(key).toString(), 2));
+			builder.append(Utilities.tabBlock(faces.get(key).toString(), 2));
 		}
 		return builder.toString();
 	}
