+++ src/main/java/useless/dragonfly/model/block/data/PositionData.java
@@ -1,7 +1,8 @@
 package useless.dragonfly.model.block.data;
 
 import com.google.gson.annotations.SerializedName;
-import useless.dragonfly.DragonFly;
+
+import java.util.Arrays;
 
 public class PositionData {
 	@SerializedName("rotation")
@@ -11,6 +12,9 @@ public class PositionData {
 	@SerializedName("scale")
 	public double[] scale;
 	public String toString(){
-		return DragonFly.writeFields(PositionData.class);
+		return
+			"rotation: " + Arrays.toString(rotation) + "\n" +
+			"translation: " + Arrays.toString(translation) + "\n" +
+			"scale: " + Arrays.toString(scale) + "\n";
 	}
 }
