+++ src/main/java/useless/dragonfly/utilities/Utilities.java
@@ -1,0 +1,36 @@
+package useless.dragonfly.utilities;
+
+import java.lang.reflect.Field;
+
+public class Utilities {
+	public static final float COMPARE_CONST = 0.001f;
+	public static String writeFields(Class clazz){
+		Field[] fields = clazz.getFields();
+		StringBuilder builder = new StringBuilder();
+		for (int i = 0; i < fields.length; i++)  {
+			Field field = fields[i];
+			builder.append(field.getName()).append(": ").append(field.getClass().cast(field));
+			if (i < fields.length-1){
+				builder.append("\n");
+			}
+		}
+		return builder.toString();
+	}
+	public static String tabBlock(String string, int tabAmount){
+		StringBuilder builder = new StringBuilder();
+		for (String line: string.split("\n")) {
+			builder.append(tabString(tabAmount)).append(line).append("\n");
+		}
+		return builder.toString();
+	}
+	public static String tabString(int tabAmount){
+		StringBuilder builder = new StringBuilder();
+		for (int i = 0; i < tabAmount; i++) {
+			builder.append("\t");
+		}
+		return builder.toString();
+	}
+	public static boolean equalFloats(float a, float b){
+		return Math.abs(Float.compare(a, b)) < COMPARE_CONST;
+	}
+}
