+++ src/main/resources/dragonfly.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "useless.dragonfly.mixin",
+  "package": "useless.dragonfly.mixins.mixin",
   "compatibilityLevel": "JAVA_8",
   "mixins": [
     "MinecraftMixin",
